"Cleaned up" the WCFService solution's project files, and added missing DLLs and removed extraneous DLLS. Everything builds again and WCF works again. Yay!
There is a warning for two projects that they have an Npgsql conflict, since they are using a separate, slightly older DLL from the lib folder as a reference instead of a project reference to the Npgsql project (which is probably referenced by other references, producing the conflict). I'm assuming that's on purpose. But this should be fine regardless, since the versions are specifically listed in the references.

GaPSLabsWCFService has two DLLs that are copied to the output folder instead of them being a reference. This also looks to be because of version mismatches and I'm assuming it's therefore on purpose.