Got the "OSM Populate database" project working. It depends on OsmSharp 4 which in turn depends on Npgsql 2.0.12.0 and all these projects are built on and/or provide Npgsql 2.1.1.0. This leads to database population attempts ending in an unhandled exception about assembly reference issues.
For some reason the initial commit didn't work either, which leads me to wonder how exactly it was done originally. Nevertheless, I managed to scoop a 2.0.12.0 DLL of off someone's random GitHub project where it was versioned, and used an App.config assembly version redirect to "downgrade" all 2.1.1.0 Npgsql references to 2.0.12.0. With the DLL replaced it now works. It looks like only "OSM Populate database" was affected, and all other projects still use 2.1.1.0 as usual.

I also added a redirect the other way around to the "OSMParserTestApplication" project to get rid of that warning I was talking about in the previous commit. Since we're doing that now anyways...

I added the "GaPSLabsWCFservice" project to the list of projects to build in the SCFService solution so that it builds when you build the solution. It was kind of invisibly not being built and then IIS was just like "wut u want".

I have succesfully ran the FirstTest scenario described in the documentation after this.