Edited the SumoTrafficGenerator project so they refer to the Python scripts in the places they appear to be in a fresh Sumo 2.6/2.7 installation.
Added --ignore-route-errors to the Sumo invocation in start.bat so that low quality OSM sections (which occur often apparently) don't stop the simulation (the cars will teleport instead).

Moved the files from the bin/Release/files folder to the files folder and set them to copy on build. The end effect is the same when you run SumoTrafficGenerator, but now you can release or debug, as well as clean or rebuild, etc.