Addressed several causes of poor performance.
While running a simulation/scene, potentially a lot of pedestrians and public transits are spawned, often at the same time. This stacks slow operations in a single frame, causing regularly repeating noticeable drops in framerate (visible stutter).

Causes I've addressed:
- Spawning a pedestrian performed several FindObjectOfType() calls at varying stages of nesting. FindObjectOfType() is a slow operation and hundreds of these at the same time eat a lot of processor power. Since they are all searching for the same thing, this has been resolved by using a higher level cache as a parameter. FindObjectOfType() is normally only run in Awake() calls due to its heavy operation, but even though this was also true for pedestrians, dynamically spawned objecst (pedestrians) still call their Awake() during spawn time (that is, while the game is running).
- Spawning pedestrians causes them to look for an optimal itinerary, which was implemented by sorting an array and using the itinerary at the first index. This has been replaced with a simple minimum search. Additionally one of the sorting conditions was based on a total travel time, which was recalculated for every invocation of the comparison delegate. This has been replaced with a lazy cache since the values that the total travel time is based on are written only once.
- Spawning a public transit performed several FindObjectOfType() calls as well, which has been resolved in the same manner.
- Spawning public transit assigns them a new material. Creating a material is slow, so it has been replaced with a cached version (this also reuses the material).

These things caused the two large spikes I encountered during profiling. There is one more spike, in LineController.getQueingHeadCount(). This is caused by the garbage collector, which apparently has a huge hard-on for this method. Garbage collectors will garbage collect, so I left that one alone.

The performance seems pretty good now. The TU Delft scene used to have an average of 2.5 FPS (as an average of smooth frames and stutter frames), now it has a nice comfy 100 FPS all around. When you chuck up the speed to x5, and the bulk of the pedestrians spawn, it drops to 20 FPs. This is probably doable, or maybe less pedestrians should be spawned. Delft is a busy city though! ;)

Note that if you spawn a pedestrian or a public transit in the code now, you have to provide these cached values (see their Awake() methods). I'm looking at you, historical data team ;)