Added a middle mouse camera dragging function.
The camera control code seems to contain an official mouse movement implementation using the left mouse button. It performs raycasts against a ground plane in layer 9. The reason it doesn't work in any of the scenes we have is because there is nothing in layer 9.

I added my simple implementation (which assumes a ground plane of Y = 0) under middle mouse button instead, so as not to mess with the official implementation. It's also not guaranteed to work well with the rest of the control scheme. Essentially, I implemented it for development purposes, to aid and speed up testing and such.