Used a centered anchor for the camera feed module. I also fixed an issue where nothing in the scene was clickable when the camera feed module was in the scene. I had to update the prefab as well as the script for that, and as a result the camera's tranform settings are now TU Delft scene based (since I could not apply only part of the prefab). I think this might ruin Furkan's test scene :/
The issue was that the second set of three cameras weren't disabled and that mucked with Unity since they did not have a render texture assigned. I disabled them in the prefab as well as in the script. Since this also fixes the issue where the camera tries to take over, I set them all back to Display 1 (since it is ignored if they have a render texture, or are disabled).