I was wrong. Unity just didn't give me the errors until I had already pushed.
The maximum amount of pedestrians with heatmaps that we can support for DirectX 9 (SM3) with the current implementation/technique is 87.

-------------

Revert "I think I fixed the heatmap shader for Shader Model 3 (DirectX 9) as it throws no more errors, but I can't test for it. Also note that building now takes ages since it has to unroll all the loops for SM3. It also appears heatmaps kinda jump now, not sure if that is normal or a result from this."

This reverts commit c9927e8df8a8fa863b4ad30f5d6cfe1e31c456d4.