Gave the TU Delft scene a new look, as requested. To be honest, I think it looks a lot less noisy, and a lot better due to the lack of weirdly scaled textures. The heatmap is definitely stands out more now, which I suspect is what Jayanth was going for.
I also disabled the ground plate to let the dark background do its thing instead, and I disabled the Mekelpark area since it seems like it's incorrectly shaped and causes a lot of Z-fighting. Pedestrians will still walk over this area, so it's purely a visual thing.

I used to previously created tool to replace the materials in the scene with the materials in this commit. As per the tool, the original materials are stored in the game objects, and can be restored if necessary.