Moved the camera render textures to the camera feeds folder. The camera feeds now use AA like the scene. I resized the textures to 200x200 to fit the camera feeds without distortion. This makes them non-power-of-two but that should be fine.
Removed all the quality levels so users can only play at the ProtoWorld level. This quality level forces AA to be on.

Perhaps forcing AA to be on is not such a good idea, but it does help a lot. We do know the target machine can handle it as well. Plus the render textures have a specific separate setting so it's the only (quick) way to have it all the same.