Added the bus and walk routes that wouldn't save yesterday. They saved today (???). I kept the walking distance at 1, since the walk routes that it would generate have already been added to the scene manually.
Fixed various smaller things, and solved the z-fighting issues by increasing the near plane (didn't realize that would make such a huge difference). I was able to move the Tiber under the roads :D

I intend to fix the area z-fighting too but first I need to replace the areas I deleted (not a good idea, should've disabled them instead). For this I first need to commit (or stash, but eh).

I also moved things around a teensy bit in the UI, and switched the heatmap and log bits so the disable simulation things button doesn't leave a gap.

Furkan fixed the heatmap and MMV scrollbar stuff. I resized, moved, and reanchored the MMV stuff to the top right of the screen like the KPIs. Note: It's now set for Full HD resolution. (Tip: you can specify the resolution you want to see in the Game tab. It scales down the view if it doesn't fit.)

I changed the camera feeds so they are under the same angle as the camera, fixing the issue with the angles. I reduced their far plane and increased their near plane to get rid of the render artifacts in the rendertextures.

I added all the simulation things to the ConfigUI toggler.

I think that's it.