Deleted areas have been restored and the entire group has been moved down 0.5 units so they render under the roads. I restored the Tiber's Y coordinate to 0, and also moved the entire water group (this includes the Tiber) down by 0.5 units so they render under the roads. The new camera settings from the previous commit allow this, z-fighting wise. Super sweet.
I reenabled the logo and put it above the KPIs. It's in the navigation prefab too. I made it fit better in the UI.

I slightly jiggled the KPIs around so the preview looks like the real view, and the charts are no longer offset in the real view. This is in the prefab too.

I disabled the logspam from the boundary stuff, as well as the MMV scrollbar. It works, so I doubt it's still necessary XD

This should be it for the release! Woohoo!