Finished the arbitrary heatmap limit setting. In the end, it was rather straightforward, and produced very uncomplex code, but we are tied to DirectX11 now. Not that this matters, the DirectX9 usability was extremely minimal, so... But Unity no longer throws errors while building now given the new pragma instruction, it will only compile the DX11 shader. I also updated the various heat producers to no longer use the 50% chance thing (since it's no longer necessary) and added comments clarifying the performance optimizations regarding heatmap setting. I also upped the count maximum for the Rome scene.
The heatmap values are now passed through a ComputeBuffer to the vertex shader. This has no limit nor requires any fancy texture tricks. It's simply a DirectX11 feature available in Unity. The trick was to learn of its existance. The shader also no longer does more work than necessary.

Note that while this provides a scalable solution, there is a cost to the compute buffer due to its upload, and the cost increases as its size (the number of data points in the scene) increases. This is noticeable on Rome on less powerful machines. Performance is currently warped by the giant garbage collector calls though, so it's hard to really observe it correctly. Still, I see regular spikes in the profiler... however, it doesn't seem to line up with the compute buffer per say (the spike interval appears slower than the heatmap refresh, and changing the wait time doesn't seem to change the spike interval), so I'm not entirely sure that is the problem.