This'll have to do for now. Haifa scene is beautiful. Venice scene... not so much. Don't get the wrong, the island is awesome, but the flow visualisation is terrible at the moment.
Both scenes use the same flow visualization at the moment (though with slight differences), but the Venice scene just needs a whole other kind. One that involves the roads and where you can click on PoIs to selectively enable flow visualizations. I think.

The BlueToothOverlay is now the FlowVisualizationModule, but the code is very ad hoc and messy. It basically only works for Haifa and Venice. It's not part of the ProtoWorld Editor menu.