Made a quick-and-dirty MatSIM plans file score parser.
It works similar to the other one from the command line.

It creates a scores table with pid and score columns, and a row for the score of the selected plan of each person.

It doesn't load everything into memory so you don't need a lot of RAM. It has transaction batching for improved performance.