Added score display through the heatmap. It compiles but is untested. I will hopefully be able to test it tomorrow.
I also just started hacking my way through the other modules since I'm just pressed for time. I disabled the time slider since it's an UI element that's in the way and that I need to test and I have to make sacrifices.

I hacked a container class into the heatmap and disabled everything but vehicles so I can dynamically update the score with minimal work.

Scores of personal vehicles and public transit vehicles are now pulled from the database and hopefully displayed. Personal vehicles have an ID that is the same as the person ID, so that's nice. Public transit vehicles requires a big-ass query and data structure to sum the scores depending on the persons riding it. But it's a big-ass server they're running it on, so...